backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
pydantic==2.5.2
googlemaps==4.10.0
python-dotenv==1.0.0
google-genai==0.3.0
